Cleaned the column names in full_dataset and scaled_dataset.
Ensured that all symbols are now properly paired with their respective indicators in plot labels and tooltips.

Moved the bulk to t-SNE calculation off the dashboard and into a tibble read from a .rds file.

Added an in-depth conclusion to the model applying the t-SNE plot to demonstrate the regime seperation ability of the model.

Fixed the "subscript out of bounds" error with the table in the output indicators.